+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # 工具库更新日志
 
+## [1.0.2](https://github.com/RootWater/toolbox/compare/v1.0.1...v1.0.2) (2021-02-02)
+
+
+### Bug Fixes
+
+* **package.json, .gitignore:** 修复发布问题 ([7d1d5ef](https://github.com/RootWater/toolbox/commit/7d1d5ef50db8373127e5c2c55d2c8bcaea82b378))
+
+
+### Features
+
+* **object-tool, string-tool:** 新增工具函数 ([cdbb876](https://github.com/RootWater/toolbox/commit/cdbb876fc2c2f028c2bede148b4947875b1aa25b))
+
+
+
 ## [1.0.1](https://github.com/RootWater/toolbox/compare/58a028b2839b6609f591bc57eb55b5b1f5b0ecbf...v1.0.1) (2021-02-01)